R-Daneel: Added conda shell setup block back into ~/.profile.
Needed to use NCL today, and the quickest way to install it turns out
to be through conda (this is actually the preferred way to install it,
per the official instructions). To "re-enable" my existing conda
install, I had to add this conda-shell setup chunk back into my bash env.