mother-night: Added package man-db.
In order to get manpage completions working in Emacs, the manpage db
needs to be periodically updated MANUALLY by the user. In other
distros like Arch, this is normally handled by a systemd service, but
for some reason this is not done in NixOS (and I don't care to create
one myself).

See also https://nixos.wiki/wiki/Apropos

The following is an excerpt from the link above:

As of NixOS 18.09, apropos (and likely also whatis) does not find
anything when run. This is because the Apropos database is not
generated automatically.

To generate it manually (and again to update it when new software is
installed), run:

sudo mkdir -p /var/cache/man/nixos
sudo mandb