aunuu: Setup monitors earlier in boot process
The xrandrHeads attr is used to setup the screen rotation earlier in
the boot process, which allows the login screen and wallpaper to
properly rotate on the portrait-mode screen. Unfortunately, I haven't
been able to properly set the primary screen (nor the arrangement),
using this mechanism, which is necessary to properly setup
text-scaling.

As a fudge, the monitors are re-configured by xrandr after the display
manager starts, so that we can set the primary and the arrangement.